+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.17.1
+- improve handling of timeout and unknownhost exceptions on android
+
 ## 1.17.0
 - vs : 16.8.4 (xamarin.vs 16.8.000.261 ; xamarin.android 11.1.0.26 ; xamarin.ios 14.8.0.3)
 - Portable.BouncyCastle 1.8.9